WORD/requirements.txt
python-docx
pywin32
latex2mathml
lxml
